Fix issue with resuming campaign creation process
The intent to start a new activity would have threading issues. This was a result of trying to access non-static information within the onclick listener. This commit resolves the issue by instead tagging each view that the adapter creates with the proper associated campaign ID and visiblity of character race and class. For now, it's simple to go off of visibility of elements - if not visible, then they were not set and require to be set. In the future where the steps can change, it would be better to create a step counter and increment each time an activity in the campaign creation process closes (upon a successful campaign update). Each activity to resume by will be associated with a step number, which can be outlined ina  fixed format by using resources like strings.xml