Remove unneeded test for background randomization
The test to check if a default single background wil be chosen again is
unnecessary and does not contribute towards testing behavior. Additionally, the Array filter method is not used anymore and a stub relied on it for the test. Backgrounds can now be reselected at random after recent changes, but with possibly different decisions. While the decision change can be tested, there is no comparison and the same decisions could be chosen (unlikely, but possible). This may be remodeled from a form's perspective to generate the possible form to fill out in advance. It can be done as a radio button list that populates on background selection within the same form instead of having separate forms per decision.